# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DownloadRequest {
  /** The YouTube/SoundCloud/etc URL to download audio from */
  url: string;
}

export interface TrackInfo {
  title: string;
  uploader: string;
  duration?: number | null;
  thumbnail?: string | null;
  platform: string;
}

export interface DownloadResult {
  /** Token to use with GET /download/file/{token} */
  token: string;
  title: string;
  filename: string;
}

export interface PlaylistTrack {
  index: number;
  id: string;
  url: string;
  title: string;
  duration?: number | null;
  thumbnail?: string | null;
}

export interface PlaylistInfo {
  title: string;
  trackCount: number;
  tracks: PlaylistTrack[];
}

export interface ErrorResponse {
  error: string;
}