# lib/api-zod/src/generated/types/playlistInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlaylistTrack } from "./playlistTrack";

export interface PlaylistInfo {
  title: string;
  trackCount: number;
  tracks: PlaylistTrack[];
}