# lib/api-zod/src/generated/types/playlistTrack.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlaylistTrack {
  index: number;
  id: string;
  url: string;
  title: string;
  duration?: number | null;
  thumbnail?: string | null;
}